Fix some warnings in MediaBrowser.Controller
